+++ app/api/dds/v1/software_agents_api.rb
@@ -6,12 +6,13 @@ module DDS
         named 'list software agents'
         failure [
           [200, 'Success'],
-          [401, 'Unauthorized']
+          [401, 'Unauthorized'],
+          [403, 'Forbidden (software_agent restrictured)']
         ]
       end
       get '/software_agents', root: 'results' do
         authenticate!
-        SoftwareAgent.where(is_deleted: false)
+        policy_scope(SoftwareAgent).where(is_deleted: false)
       end
 
       desc 'Create a software agent' do
@@ -21,7 +22,8 @@ module DDS
           [200, 'This will never actually happen'],
           [201, 'Created Successfully'],
           [400, 'Software agent requires a name'],
-          [401, 'Unauthorized']
+          [401, 'Unauthorized'],
+          [403, 'Forbidden (software_agent restrictured)']
         ]
       end
       params do
@@ -38,6 +40,7 @@ module DDS
           repo_url: software_agent_params[:repo_url],
           creator: current_user
         })
+        authorize software_agent, :create?
         software_agent.build_api_key(key: SecureRandom.hex)
         Audited.audit_class.as_user(current_user) do
           if software_agent.save
@@ -55,7 +58,7 @@ module DDS
         failure [
           [200, 'Success'],
           [401, 'Unauthorized'],
-          [403, 'Forbidden'],
+          [403, 'Forbidden (software_agent restricted)'],
           [404, 'Software Agent Does not Exist']
         ]
       end
@@ -64,7 +67,65 @@ module DDS
       end
       get '/software_agents/:id', root: false do
         authenticate!
-        SoftwareAgent.find(params[:id])
+        sa = SoftwareAgent.find(params[:id])
+        authorize sa, :show?
+        sa
+      end
+
+      desc 'Update Software Agent' do
+        detail 'Updates the software agent details for a given UUID.'
+        named 'update software agent'
+        failure [
+          [200, 'Success'],
+          [401, 'Unauthorized'],
+          [403, 'Forbidden (software_agent restricted)'],
+          [400, 'Validation Error'],
+          [404, 'Software Agent Does not Exist']
+        ]
+      end
+      params do
+        requires :id, type: String, desc: 'Software Agent UUID'
+        optional :name, type: String, desc: 'The Name of the Software Agent'
+        optional :description, type: String, desc: 'The Description of the Software Agent'
+        optional :repo_url, type: String, desc: 'The Repo url of the Software Agent'
+      end
+      put '/software_agents/:id', root: false do
+        authenticate!
+        software_agent_params = declared(params, include_missing: false)
+        software_agent = hide_logically_deleted SoftwareAgent.find(params[:id])
+        authorize software_agent, :update?
+        Audited.audit_class.as_user(current_user) do
+          if software_agent.update(software_agent_params)
+            annotate_audits [software_agent.audits.last]
+            software_agent
+          else
+            validation_error!(software_agent)
+          end
+        end
+      end
+
+      desc 'Delete a Software Agent' do
+        detail 'Marks a software agent as being deleted.'
+        named 'delete software agent'
+        failure [
+          [204, 'Successfully Deleted'],
+          [401, 'Unauthorized'],
+          [403, 'Forbidden (software_agent restricted)'],
+          [404, 'Software Agent Does not Exist']
+        ]
+      end
+      params do
+        requires :id, type: String, desc: 'Software Agent UUID'
+      end
+      delete '/software_agents/:id', root: false do
+        authenticate!
+        software_agent = hide_logically_deleted SoftwareAgent.find(params[:id])
+        authorize software_agent, :destroy?
+        Audited.audit_class.as_user(current_user) do
+          software_agent.update(is_deleted: true)
+          annotate_audits [software_agent.audits.last]
+        end
+        body false
       end
 
       desc 'Re-generate software agent API key' do
@@ -73,7 +134,7 @@ module DDS
         failure [
           [200, 'Success'],
           [401, 'Unauthorized'],
-          [403, 'Forbidden'],
+          [403, 'Forbidden (software_agent restricted)'],
           [404, 'Software Agent Does not Exist']
         ]
       end
@@ -83,15 +144,74 @@ module DDS
       put '/software_agents/:id/api_key', serializer: ApiKeySerializer do
         authenticate!
         software_agent = SoftwareAgent.find(params[:id])
+        authorize software_agent.api_key, :update?
         Audited.audit_class.as_user(current_user) do
           ApiKey.transaction do
+            original_api_key_id = software_agent.api_key.id
             software_agent.api_key.destroy!
+            delete_audit = Audited.audit_class.where(auditable_id: original_api_key_id).last
             software_agent.build_api_key(key: SecureRandom.hex)
             software_agent.save
-            annotate_audits [software_agent.api_key.audits.last]
+            annotate_audits [software_agent.api_key.audits.last, delete_audit]
           end
         end
         software_agent.api_key
+      end
+      desc 'View software agent API key' do
+        detail 'View software_agent api_key'
+        named 'view software_agent api_key'
+        failure [
+          [200, 'Success'],
+          [401, 'Unauthorized'],
+          [403, 'Forbidden (software_agent restricted)'],
+          [404, 'Software Agent Does not Exist']
+        ]
+      end
+      params do
+        requires :id, type: String, desc: 'Software agent UUID'
+      end
+      get '/software_agents/:id/api_key', serializer: ApiKeySerializer do
+        authenticate!
+        ak = SoftwareAgent.find(params[:id]).api_key
+        authorize ak, :show?
+        ak
+      end
+
+      desc 'Get software agent access token'do
+        detail 'Get software agent access token'
+        named 'get software_agent access token'
+        failure [
+          [200, 'This will never happen'],
+          [201, 'Success'],
+          [400, 'Missing Required Keys'],
+          [404, 'Software Agent or User Does not Exist']
+        ]
+      end
+      params do
+        requires :agent_key, type: String, desc: 'Software agent secret key'
+        requires :user_key, type: String, desc: 'User secret key'
+      end
+      rescue_from Grape::Exceptions::ValidationErrors do |e|
+        error_json = {
+          "error" => 400,
+          "reason" => 'missing key or keys',
+          "suggestion" => 'api_key and user_key are required'
+        }
+        error!(error_json, 400)
+      end
+      rescue_from ActiveRecord::RecordNotFound do |e|
+        error_json = {
+          "error" => 404,
+          "reason" => "invalid key",
+          "suggestion" => "ensure both keys are valid"
+        }
+        error!(error_json, 404)
+      end
+      post '/software_agents/api_token', serializer: ApiTokenSerializer do
+        user_key = ApiKey.where(key: params[:user_key]).joins(:user).take!
+        software_key = ApiKey.where(key: params[:agent_key]).joins(:software_agent).take!
+        user_key.user.update_attribute(:last_login_at, DateTime.now)
+        ApiToken.new(user: user_key.user, software_agent: software_key.software_agent)
       end
     end
   end
